moving route building to http for clarity
