--- Sources/Routing/RouteBuilder+RouteCollection.swift
@@ -1,9 +1,0 @@
-//extension RouteBuilder {
-//    /**
-//        Adds a RouteCollection to the RouteBuilder
-//        by invoking the `.build` method.
-//    */
-//    public func collection<C: RouteCollection>(_ collection: C) where C.Wrapped == Value {
-//        collection.build(self)
-//    }
-//}
